Sync Alpha To Beta

+++ RockWeb/Blocks/Communication/CommunicationEntryWizard.ascx.cs
@@ -219,6 +219,9 @@ namespace RockWeb.Blocks.Communication
             if ( communication == null )
             {
                 communication = new Rock.Model.Communication() { Status = CommunicationStatus.Transient };
+                communication.CreatedByPersonAlias = this.CurrentPersonAlias;
+                communication.CreatedByPersonAliasId = this.CurrentPersonAliasId;
+                communication.SenderPersonAlias = this.CurrentPersonAlias;
                 communication.SenderPersonAliasId = CurrentPersonAliasId;
                 communication.EnabledLavaCommands = GetAttributeValue( "EnabledLavaCommands" );
 
@@ -233,7 +236,8 @@ namespace RockWeb.Blocks.Communication
             if ( editableStatuses.Contains( communication.Status ) || ( communication.Status == CommunicationStatus.PendingApproval && editingApproved ) )
             {
                 // Make sure they are authorized to view
-                if ( !communication.IsAuthorized( Rock.Security.Authorization.EDIT, CurrentPerson ) )
+                bool isCreator = ( communication.CreatedByPersonAlias != null && CurrentPersonId.HasValue && communication.CreatedByPersonAlias.PersonId == CurrentPersonId.Value );
+                if ( !communication.IsAuthorized( Rock.Security.Authorization.EDIT, CurrentPerson ) && !isCreator )
                 {
                     // not authorized, so hide this block
                     this.Visible = false;
@@ -982,7 +986,15 @@ namespace RockWeb.Blocks.Communication
         {
             var rockContext = new RockContext();
 
-            var templateQuery = new CommunicationTemplateService( rockContext ).Queryable().Where( a => a.IsActive ).OrderBy( a => a.Name );
+            var templateQuery = new CommunicationTemplateService( rockContext ).Queryable().Where( a => a.IsActive );
+
+            int? categoryId = cpCommunicationTemplate.SelectedValue.AsIntegerOrNull();
+            if (categoryId.HasValue && categoryId > 0)
+            {
+                templateQuery = templateQuery.Where( a => a.CategoryId == categoryId );
+            }
+
+            templateQuery = templateQuery.OrderBy( a => a.Name );
 
             // get list of templates that the current user is authorized to View
             IEnumerable<CommunicationTemplate> templateList = templateQuery.AsNoTracking().ToList().Where( a => a.IsAuthorized( Rock.Security.Authorization.VIEW, this.CurrentPerson ) ).ToList();
@@ -1141,6 +1153,16 @@ namespace RockWeb.Blocks.Communication
             }
         }
 
+        /// <summary>
+        /// Handles the SelectItem event of the cpCommunicationTemplate control.
+        /// </summary>
+        /// <param name="sender">The source of the event.</param>
+        /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
+        protected void cpCommunicationTemplate_SelectItem( object sender, EventArgs e )
+        {
+            BindTemplatePicker();
+        }
+
         #endregion Template Selection
 
         #region Email Editor
@@ -1217,6 +1239,15 @@ namespace RockWeb.Blocks.Communication
                 return;
             }
 
+            try
+            {
+                CleanupOrphanedTestCommunications();
+            }
+            catch( Exception ex)
+            {
+                this.LogException( ex );
+            }
+
             Rock.Model.Communication communication = UpdateCommunication( new RockContext() );
 
             if ( communication != null )
@@ -1225,9 +1256,12 @@ namespace RockWeb.Blocks.Communication
                 {
                     // Using a new context (so that changes in the UpdateCommunication() are not persisted )
                     int? testPersonId = null;
+                    Rock.Model.Communication testCommunication = null;
+                    CommunicationService communicationService = null;
+
                     try
                     {
-                        var testCommunication = communication.Clone( false );
+                        testCommunication = communication.Clone( false );
                         testCommunication.Id = 0;
                         testCommunication.Guid = Guid.Empty;
                         testCommunication.EnabledLavaCommands = GetAttributeValue( "EnabledLavaCommands" );
@@ -1277,7 +1311,10 @@ namespace RockWeb.Blocks.Communication
 
                         testPerson.ForeignGuid = null;
                         testPerson.ForeignId = null;
-                        testPerson.ForeignKey = null;
+
+                        // Just in case it doesn't get cleaned up, mark it so that we can try to clean it up next time
+                        testPerson.ForeignKey = "_ForTestCommunication_";
+
                         var personService = new PersonService( rockContext );
                         personService.Add( testPerson );
                         rockContext.SaveChanges();
@@ -1293,7 +1330,7 @@ namespace RockWeb.Blocks.Communication
 
                         testCommunication.Recipients.Add( testRecipient );
 
-                        var communicationService = new CommunicationService( rockContext );
+                        communicationService = new CommunicationService( rockContext );
                         communicationService.Add( testCommunication );
                         rockContext.SaveChanges();
 
@@ -1320,26 +1357,66 @@ namespace RockWeb.Blocks.Communication
 
                         nbResult.Dismissable = true;
                         nbResult.Visible = true;
-
-                        communicationService.Delete( testCommunication );
-                        rockContext.SaveChanges();
                     }
                     finally
                     {
-                        // make sure we delete the Person record we created to send the test email
+                        // make sure we delete the test communication record we created to send the test 
+                        if ( communicationService != null && testCommunication != null )
+                        {
+                            communicationService.Delete( testCommunication );
+                            rockContext.SaveChanges();
+                        }
+
+                        // make sure we delete the Person record we created to send the test 
                         using ( var deleteRockContext = new RockContext() )
                         {
+                            var personToDeleteService = new PersonService( deleteRockContext );
+                            var personAliasToDeleteService = new PersonAliasService( deleteRockContext );
+
                             if ( testPersonId.HasValue )
                             {
-                                var personToDeleteService = new PersonService( deleteRockContext );
-                                var personAliasToDeleteService = new PersonAliasService( deleteRockContext );
                                 var personToDelete = personToDeleteService.Get( testPersonId.Value );
                                 personAliasToDeleteService.DeleteRange( personToDelete.Aliases );
                                 personToDelete.Aliases = null;
                                 personToDeleteService.Delete( personToDelete );
                                 deleteRockContext.SaveChanges();
                             }
+}
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Shouldn't happen, but just in case...
+        /// this will cleanup any orphaned test communications artifacts
+        /// </summary>
+        private void CleanupOrphanedTestCommunications()
+        {
+            using ( RockContext deleteRockContext = new RockContext() )
+            {
+                PersonService personToDeleteService = new PersonService( deleteRockContext );
+                PersonAliasService personAliasToDeleteService = new PersonAliasService( deleteRockContext );
+                
+                // check for any test communication artifacts that didn't clean up correctly
+                var forTestCommunicationPersonQry = personToDeleteService.Queryable().Where( a => a.ForeignKey == "_ForTestCommunication_" );
+                if ( forTestCommunicationPersonQry.Any() )
+                {
+                    var communicationToDeleteService = new Rock.Model.CommunicationService( deleteRockContext );
+                    var communicationToDeleteQry = communicationToDeleteService.Queryable().Where( a => a.Recipients.Any( r => forTestCommunicationPersonQry.Any( p => p.Id == r.PersonAlias.PersonId ) ) );
+                    foreach ( var testCommunicationToDelete in communicationToDeleteQry.ToList() )
+                    {
+                        communicationToDeleteService.Delete( testCommunicationToDelete );
+                    }
+
+                    foreach ( var personToDelete in forTestCommunicationPersonQry )
+                    {
+                        foreach ( var personToDeleteAlias in personToDelete.Aliases.ToList() )
+                        {
+                            personAliasToDeleteService.Delete( personToDeleteAlias );
                         }
+
+                        personToDeleteService.Delete( personToDelete );
                     }
                 }
             }
@@ -1686,7 +1763,9 @@ namespace RockWeb.Blocks.Communication
                         }
                     }
 
-                    imgSMSImageAttachment.ImageUrl = virtualThumbnailFilePath;
+                    string publicAppRoot = GlobalAttributesCache.Read().GetValue( "PublicApplicationRoot" ).EnsureTrailingForwardslash();
+                    imgSMSImageAttachment.ImageUrl = virtualThumbnailFilePath.Replace( "~/", publicAppRoot );
+                    divAttachmentLoadError.InnerText = "Unable to load preview icon from " + imgSMSImageAttachment.ImageUrl;
                     imgSMSImageAttachment.Visible = true;
                     imgSMSImageAttachment.Width = new Unit( 10, UnitType.Percentage );
                 }
@@ -1792,6 +1871,15 @@ namespace RockWeb.Blocks.Communication
         /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
         protected void btnSMSSendTest_Click( object sender, EventArgs e )
         {
+            if (ddlSMSFrom.SelectedValue.IsNullOrWhiteSpace())
+            {
+                nbSMSTestResult.NotificationBoxType = NotificationBoxType.Danger;
+                nbSMSTestResult.Text = "A 'From' number must be specified.";
+                nbSMSTestResult.Dismissable = true;
+                nbSMSTestResult.Visible = true;
+                return;
+            }
+
             SendTestCommunication( EntityTypeCache.Read( Rock.SystemGuid.EntityType.COMMUNICATION_MEDIUM_SMS.AsGuid() ).Id, nbSMSTestResult );
         }
 
@@ -2240,6 +2328,8 @@ sendCountTerm.PluralizeIf( sendCount != 1 ) );
         }
 
         #endregion
+
+
 
         
     }
